not working

+++ src/pt/diogo/jwallet/ICurrency.java
@@ -2,5 +2,5 @@ package pt.diogo.jwallet;
 
 public interface ICurrency {
 	String getLabel();
-	Long getValue();
+	int getValue();
 }
